pyproject.toml
[project]
name = "akong-agent-harness"
version = "0.0.1"
description = "通用 agent harness · LLM 周围的软件基础设施 · workspace + memory + tools + runtime 4 件套 · 跨平台"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai==2.36.0",
    "httpx==0.28.1",
    "python-dateutil==2.9.0.post0",
]

[dependency-groups]
dev = [
    "pytest==9.0.3",
    "pytest-mock==3.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/akong_agent_harness"]

[tool.pytest.ini_options]
testpaths = ["tests"]
